UI toolkit integration (WIP)

+++ Assets/Scripts/Settings/AudioController.cs
@@ -6,7 +6,7 @@ namespace Settings
 {
     public class AudioController : MonoBehaviour
     {
-        public event Action<float, float, float> OnSettingsLoaded;
+        public event Action<int, int, int> OnSettingsLoaded;
         [SerializeField] private AudioMixer audioMixer;
         private const string MASTER_VOLUME = "MasterVolume";
         private const string MUSIC_VOLUME = "MusicVolume";
@@ -14,17 +14,17 @@ namespace Settings
 
         private void Awake() => Load();
 
-        public void SetMasterVolume(float volume) => audioMixer.SetFloat(MASTER_VOLUME, volume);
+        public void SetMasterVolume(int volume) => audioMixer.SetFloat(MASTER_VOLUME, volume);
 
-        public void SetMusicVolume(float volume) => audioMixer.SetFloat(MUSIC_VOLUME, volume);
+        public void SetMusicVolume(int volume) => audioMixer.SetFloat(MUSIC_VOLUME, volume);
 
-        public void SetSfxVolume(float volume) => audioMixer.SetFloat(SFX_VOLUME, volume);
+        public void SetSfxVolume(int volume) => audioMixer.SetFloat(SFX_VOLUME, volume);
 
         private void Load()
         {
-            var masterVolume = PlayerPrefs.GetFloat(MASTER_VOLUME, -50f);
-            var musicVolume = PlayerPrefs.GetFloat(MUSIC_VOLUME, -50f);
-            var sfxVolume = PlayerPrefs.GetFloat(SFX_VOLUME, -50f);
+            var masterVolume = PlayerPrefs.GetInt(MASTER_VOLUME, -50);
+            var musicVolume = PlayerPrefs.GetInt(MUSIC_VOLUME, -50);
+            var sfxVolume = PlayerPrefs.GetInt(SFX_VOLUME, -50);
             OnSettingsLoaded?.Invoke(masterVolume, musicVolume, sfxVolume);
         }
